requirements.txt
fastapi
uvicorn
python-dotenv
requests
aiohttp
pydantic
colorama
langchain-groq
